float64 size and sizeleft

+++ radarr/type.go
@@ -229,9 +229,9 @@ type QueueRecord struct {
 	Languages               []*starr.Value         `json:"languages"`
 	Quality                 *starr.Quality         `json:"quality"`
 	CustomFormats           []interface{}          `json:"customFormats"` // probably []int64
-	Size                    int64                  `json:"size"`
+	Size                    float64                `json:"size"`
 	Title                   string                 `json:"title"`
-	Sizeleft                int64                  `json:"sizeleft"`
+	Sizeleft                float64                `json:"sizeleft"`
 	Timeleft                string                 `json:"timeleft"`
 	EstimatedCompletionTime time.Time              `json:"estimatedCompletionTime"`
 	Status                  string                 `json:"status"`
